Do not set promptdivs if in tmux

+++ .profile
@@ -4,16 +4,19 @@
 [ -z "$XDG_DATA_HOME" ] && export XDG_DATA_HOME="$HOME/.local/share"
 [ -z "$XDG_CACHE_HOME" ] && export XDG_CACHE_HOME="$HOME/.cache"
 
-if [ "$TERM" = "screen" ]; then
-	touch "$XDG_RUNTIME_DIR/androidterm"
-else
-	rm "$XDG_RUNTIME_DIR/androidterm" 2>/dev/null
+if [ -z "$TMUX" ]; then
+	if [ "$TERM" = "screen" ]; then
+		touch "$XDG_RUNTIME_DIR/androidterm"
+	else
+		rm "$XDG_RUNTIME_DIR/androidterm" 2>/dev/null
+	fi
 fi
 
 for profile in $XDG_CONFIG_HOME/profile.d/*; do
 	. "$profile"
 done
 unset profile
+
 [ -f ~/.bashrc ] && . ~/.bashrc
 
 if [ "$(fgconsole 2>/dev/null)" = "1" ]; then
